Struct leptess::LepTess

pub struct LepTess { /* private fields */ }

High level wrapper for Tesseract and Leptonica

Examples

Full page OCR

let mut lt = leptess::LepTess::new(Some("./tests/tessdata"), "eng").unwrap();
lt.set_image("./tests/di.png");
println!("{}", lt.get_utf8_text().unwrap());

OCR on a specific region of the image

lt.set_rectangle(10, 10, 200, 60);
println!("{}", lt.get_utf8_text().unwrap());

Iterate bounding boxes for each recognized word

let boxes = lt.get_component_boxes(
    leptess::capi::TessPageIteratorLevel_RIL_WORD,
    true,
).unwrap();

for b in &boxes {
    println!("{:?}", b);
}

Implementations

impl LepTess

pub fn new(data_path: Option<&str>, lang: &str) -> Result<LepTess, TessInitError>

pub fn set_image(&mut self, img_uri: impl AsRef<Path>) -> Result<(), PixError>

Set image to use for OCR.

pub fn set_image_from_mem(&mut self, img: &[u8]) -> Result<(), PixError>

Set the source image from an in-memory file

Only tiff files are supported from windows. More file formats are supported from other operating systems

pub fn get_source_y_resolution(&mut self) -> i32

pub fn get_image_dimensions(&self) -> Option<(u32, u32)>

pub fn set_source_resolution(&mut self, res: i32)

Override image resolution. Can be used to suppress “Warning: Invalid resolution 0 dpi.” output.

pub fn set_fallback_source_resolution(&mut self, res: i32)

Override image resolution if not detected

pub fn recognize(&mut self) -> i32

pub fn set_rectangle(&mut self, left: i32, top: i32, width: i32, height: i32)

Restrict OCR to a specific region of the image.

pub fn set_rectangle_from_box(&mut self, b: &Box)

Restrict OCR to a specific region of the image using a leptonica Box struct.

pub fn get_utf8_text(&mut self) -> Result<String, Utf8Error>

Extract text from current selected region of the image. By default, it is the full page. But it can be changed through set_rectangle api.

Example

let mut lt = leptess::LepTess::new(None, "eng").unwrap();
lt.set_image("./tests/di.png");
println!("{}", lt.get_utf8_text().unwrap());

pub fn get_hocr_text(&mut self, page: c_int) -> Result<String, Utf8Error>

Extract text from image as HTML with bounding box attributes.

pub fn get_alto_text(&mut self, page: c_int) -> Result<String, Utf8Error>

Extract text from image as XML-formatted string with Alto markup.

pub fn get_tsv_text(&mut self, page: c_int) -> Result<String, Utf8Error>

Extract text from image as TSV-formatted string.

pub fn get_lstm_box_text(&mut self, page: c_int) -> Result<String, Utf8Error>

Returns a box file for LSTM training from the internal data structures. Constructs coordinates in the original image - not just the rectangle.

pub fn get_word_str_box_text(&mut self, page: c_int) -> Result<String, Utf8Error>

Extract text from image as a string formatted in the same way as a Tesseract WordStr box file used in training.

pub fn mean_text_conf(&self) -> i32

pub fn get_regions(&self) -> Option<Boxa>

pub fn get_component_boxes(
&self,
level: TessPageIteratorLevel,
text_only: bool
) -> Option<Boxa>

Get the given level kind of components (block, textline, word etc.) as a leptonica-style Boxa, in reading order. If text_only is true, then only text components are returned.

Example

Get word bounding boxes

let mut lt = leptess::LepTess::new(None, "eng").unwrap();
lt.set_image("./tests/di.png");
let boxes = lt.get_component_boxes(
    leptess::capi::TessPageIteratorLevel_RIL_WORD,
    true,
).unwrap();

for b in &boxes {
    println!("{:?}", b);
}

pub fn set_variable(
&mut self,
name: Variable,
value: &str
) -> Result<(), TessSetVariableError>

Set the value of an internal Tesseract parameter.

Example

let mut lt = leptess::LepTess::new(None, "eng").unwrap();
lt.set_variable(leptess::Variable::TesseditCharBlacklist, "xyz").unwrap();