Struct leptess::LepTess

pub struct LepTess { /* fields omitted */ }

High level wrapper for Tesseract and Leptonica

Examples

Full page OCR

let mut lt = leptess::LepTess::new(Some("./tests/tessdata"), "eng").unwrap();
lt.set_image("./tests/di.png");
println!("{}", lt.get_utf8_text().unwrap());

OCR on a specific region of the image

lt.set_rectangle(&leptess::leptonica::Box::new(
    10, 10, 200, 60,
).unwrap());
println!("{}", lt.get_utf8_text().unwrap());

Iterate bounding boxes for each recognized word

let boxes = lt.get_component_boxes(
    leptess::capi::TessPageIteratorLevel_RIL_WORD,
    true,
).unwrap();

for b in &boxes {
    println!("{:?}", b.get_val());
}

Implementations

impl LepTess

pub fn new(
data_path: Option<&str>,
lang: &str
) -> Result<LepTess, TessInitError>

pub fn set_image(&mut self, img_uri: impl AsRef<Path>) -> Result<(), PixError>

Set image to use for OCR.

pub fn set_image_from_mem(&mut self, img: &[u8]) -> Result<(), PixError>

pub fn get_source_y_resolution(&mut self) -> i32

pub fn get_image_dimensions(&self) -> Option<(u32, u32)>

pub fn set_source_resolution(&mut self, res: i32)

Override image resolution. Can be used to suppress "Warning: Invalid resolution 0 dpi." output.

pub fn set_fallback_source_resolution(&mut self, res: i32)

Override image resolution if not detected

pub fn recognize(&self) -> i32

pub fn set_rectangle(&mut self, b: &Box)

Restrict OCR to a specific region of the image.

pub fn get_utf8_text(&self) -> Result<String, Utf8Error>

Extract text from current selected region of the image. By default, it is the full page. But it can be changed through set_rectangle api.

Example

let mut lt = leptess::LepTess::new(None, "eng").unwrap();
lt.set_image("./tests/di.png");
println!("{}", lt.get_utf8_text().unwrap());

pub fn get_hocr_text(&self, page: c_int) -> Result<String, Utf8Error>

Extract text from image as HTML with bounding box attributes.

pub fn mean_text_conf(&self) -> i32

pub fn get_regions(&self) -> Option<Boxa>

pub fn get_component_boxes(
&self,
level: TessPageIteratorLevel,
text_only: bool
) -> Option<Boxa>

Get the given level kind of components (block, textline, word etc.) as a leptonica-style Boxa, in reading order.If text_only is true, then only text components are returned.

Example

Get word bounding boxes

let mut lt = leptess::LepTess::new(None, "eng").unwrap();
lt.set_image("./tests/di.png");
let boxes = lt.get_component_boxes(
    leptess::capi::TessPageIteratorLevel_RIL_WORD,
    true,
).unwrap();

for b in &boxes {
    println!("{:?}", b.get_val());
}