Struct leptess::tesseract::TessApi

pub struct TessApi {
    pub raw: TessBaseApi,
}

Fields
raw: TessBaseApi
Implementations

impl TessApi

pub fn new<'a>(
data_path: Option<&'a str>,
lang: &'a str
) -> Result<TessApi, TessInitError>

pub fn set_image(&mut self, img: &Pix)

Provide an image for Tesseract to recognize.

set_image clears all recognition results, and sets the rectangle to the full image, so it may be followed immediately by a [Self::get_utf8_text], and it will automatically perform recognition.

pub fn get_image_dimensions(&self) -> Option<(u32, u32)>

Get the dimensions of the currently loaded image, or None if no image is loaded.

Example

let path = std::path::Path::new("tests/di.png");
let img = leptess::leptonica::pix_read(&path).unwrap();

let mut tes = leptess::tesseract::TessApi::new(Some("tests/tessdata"), "eng").unwrap();
tes.set_image(&img);

assert_eq!(tes.get_image_dimensions(), Some((442, 852)));

pub fn get_source_y_resolution(&self) -> i32

pub fn set_source_resolution(&mut self, res: i32)

Override image resolution. Can be used to suppress “Warning: Invalid resolution 0 dpi.” output.

pub fn recognize(&mut self) -> i32

pub fn set_rectangle(&mut self, b: impl AsRef<Box>)

pub fn get_utf8_text(&mut self) -> Result<String, Utf8Error>

pub fn get_hocr_text(&mut self, page: c_int) -> Result<String, Utf8Error>

pub fn get_alto_text(&mut self, page: c_int) -> Result<String, Utf8Error>

pub fn get_tsv_text(&mut self, page: c_int) -> Result<String, Utf8Error>

pub fn get_lstm_box_text(&mut self, page: c_int) -> Result<String, Utf8Error>

pub fn get_word_str_box_text(
&mut self,
page: c_int
) -> Result<String, Utf8Error>

pub fn mean_text_conf(&self) -> i32

pub fn get_regions(&self) -> Option<Boxa>

pub fn get_component_images(
&self,
level: TessPageIteratorLevel,
text_only: bool
) -> Option<Boxa>

Get the given level kind of components (block, textline, word etc.) as a leptonica-style Boxa, in reading order.If text_only is true, then only text components are returned.